+++ fiddles/angular2/fiddle-0042-DeepLinks/app/app.component.ts
@@ -1,0 +1,33 @@
+import { Component }          from '@angular/core';
+import { ROUTER_DIRECTIVES }  from '@angular/router';
+
+import { DialogService }  from './dialog.service';
+import { HeroService }    from './heroes/hero.service';
+
+@Component({
+  selector: 'app',
+  template: `
+    <h1 class="title">Component Router</h1>
+    <nav>
+      <a [routerLink]="['/crisis-center']">Crisis Center</a>
+      <a [routerLink]="['/heroes']">Heroes</a>
+      <a [routerLink]="['/crisis-center/admin']">Crisis Admin</a>
+      <a [routerLink]="['/login']">Login</a>
+    </nav>
+    <router-outlet></router-outlet>
+  `,
+  providers:  [
+    HeroService,
+    DialogService
+  ],
+  directives: [ROUTER_DIRECTIVES]
+})
+export class AppComponent {
+}
+
+
+/*
+Copyright 2016 Google Inc. All Rights Reserved.
+Use of this source code is governed by an MIT-style license that
+can be found in the LICENSE file at http://angular.io/license
+*/