+++ fiddles/angular2/fiddle-0042-DeepLinks/app/crisis-center/crisis-admin.component.ts
@@ -1,0 +1,19 @@
+import { Component }         from '@angular/core';
+import { ROUTER_DIRECTIVES } from '@angular/router';
+
+@Component({
+  template:  `
+    <h3>CRISIS ADMINISTRATION</h3>
+    <p>Manage your crises here</p>
+  `,
+  directives: [ROUTER_DIRECTIVES]
+})
+
+export class CrisisAdminComponent { }
+
+
+/*
+Copyright 2016 Google Inc. All Rights Reserved.
+Use of this source code is governed by an MIT-style license that
+can be found in the LICENSE file at http://angular.io/license
+*/