+++ fiddles/angular2/fiddle-0042-DeepLinks/app/heroes/heroes.routes.ts
@@ -1,0 +1,15 @@
+import { RouterConfig }          from '@angular/router';
+import { HeroListComponent }     from './hero-list.component';
+import { HeroDetailComponent }   from './hero-detail.component';
+
+export const HeroesRoutes: RouterConfig = [
+  { path: 'heroes',  component: HeroListComponent },
+  { path: 'hero/:id', component: HeroDetailComponent }
+];
+
+
+/*
+Copyright 2016 Google Inc. All Rights Reserved.
+Use of this source code is governed by an MIT-style license that
+can be found in the LICENSE file at http://angular.io/license
+*/