+++ fiddles/angular2/fiddle-0042-DeepLinks/app/interfaces.ts
@@ -1,0 +1,19 @@
+import { CanDeactivate } from '@angular/router';
+import { Observable }    from 'rxjs/Observable';
+
+export interface CanComponentDeactivate {
+ canDeactivate: () => boolean | Observable<boolean>;
+}
+
+export class CanDeactivateGuard implements CanDeactivate<CanComponentDeactivate> {
+  canDeactivate(component: CanComponentDeactivate): Observable<boolean> | boolean {
+    return component.canDeactivate ? component.canDeactivate() : true;
+  }
+}
+
+
+/*
+Copyright 2016 Google Inc. All Rights Reserved.
+Use of this source code is governed by an MIT-style license that
+can be found in the LICENSE file at http://angular.io/license
+*/