+++ fiddles/angular2/fiddle-0042-DeepLinks/app/main.ts
@@ -1,0 +1,16 @@
+// main entry point
+import { bootstrap }            from '@angular/platform-browser-dynamic';
+import { AppComponent }         from './app.component';
+import { APP_ROUTER_PROVIDERS } from './app.routes';
+
+bootstrap(AppComponent, [
+  APP_ROUTER_PROVIDERS
+])
+.catch(err => console.error(err));
+
+
+/*
+Copyright 2016 Google Inc. All Rights Reserved.
+Use of this source code is governed by an MIT-style license that
+can be found in the LICENSE file at http://angular.io/license
+*/