+++ fiddles/angular2/fiddle-0042-DeepLinks/app_bck/components/select/field-values.pipe.spec.ts
@@ -1,0 +1,16 @@
+import {
+  it,
+  describe,
+  expect,
+  inject,
+  beforeEachProviders
+} from '@angular/core/testing';
+import { FieldValues } from './field-values.pipe';
+
+describe('FieldValues Pipe', () => {
+  beforeEachProviders(() => [FieldValues]);
+
+  it('should transform the input', inject([FieldValues], (pipe: FieldValues) => {
+      expect(pipe.transform(true)).toBe(null);
+  }));
+});