+++ fiddles/angular2/fiddle-0042-DeepLinks/app_bck/content/content.routes.ts
@@ -1,0 +1,12 @@
+import { RouterConfig }          from '@angular/router';
+import { ContentComponent }     from './content.component';
+
+export const ContentRoutes: RouterConfig = [
+    {
+        path: '',
+        redirectTo: '/reports',
+        terminal: true
+    },
+    { path: 'reports', component: ContentComponent },
+    { path: 'reports/:id', component: ContentComponent }
+];