+++ fiddles/angular2/fiddle-0042-DeepLinks/app_bck/meta.ts
@@ -1,0 +1,7 @@
+var metadata = {
+    fiddleHeader: 'Angular 2 - CSV Reporter',
+    urls: {
+        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0040-CSVReporter'
+    },
+    consoleTag: 'H O U S E ~ f i d d l e s'
+};