+++ fiddles/angular2/fiddle-0043-ComponentRouting/app/app.component.ts
@@ -1,0 +1,17 @@
+import {Component} from '@angular/core';
+import { ROUTER_DIRECTIVES }  from '@angular/router';
+import {SidebarComponent} from './sidebar/sidebar.component';
+import {ContentComponent} from './content/content.component';
+import { DataService } from './global/data.service';
+
+
+@Component({
+    selector: 'app',
+    templateUrl: './app/app.component.html',
+    styleUrls: ['./app/app.component.css'],
+    directives: [[SidebarComponent, ContentComponent, ROUTER_DIRECTIVES]],
+    providers: [DataService]
+})
+export class AppComponent {
+
+}