+++ fiddles/angular2/fiddle-0043-ComponentRouting/app/content/content.routes.ts
@@ -1,0 +1,22 @@
+import { RouterConfig }          from '@angular/router';
+import { ContentComponent } from './content.component';
+import { ContentRouterComponent } from './content-router/content-router.component';
+
+
+export const ContentRoutes: RouterConfig = [
+    {
+        path: '',
+        redirectTo: '/content/red',
+        terminal: true
+    },
+    {
+        path: 'content',
+        component: ContentComponent,
+        children: [
+            {
+                path: ':id',
+                component: ContentRouterComponent
+            }
+        ]
+    }
+];