+++ fiddles/angular2/fiddle-0043-ComponentRouting/app/global/data.service.spec.ts
@@ -1,0 +1,18 @@
+/* tslint:disable:no-unused-variable */
+
+import {
+  beforeEach, beforeEachProviders,
+  describe, xdescribe,
+  expect, it, xit,
+  async, inject
+} from '@angular/core/testing';
+import { DataService } from './data.service';
+
+describe('Data Service', () => {
+  beforeEachProviders(() => [DataService]);
+
+  it('should ...',
+      inject([DataService], (service: DataService) => {
+    expect(service).toBeTruthy();
+  }));
+});