+++ fiddles/angular2/fiddle-0043-ComponentRouting/app/sidebar/sidebar.component.spec.ts
@@ -1,0 +1,20 @@
+/* tslint:disable:no-unused-variable */
+
+import { By }           from '@angular/platform-browser';
+import { DebugElement } from '@angular/core';
+
+import {
+  beforeEach, beforeEachProviders,
+  describe, xdescribe,
+  expect, it, xit,
+  async, inject
+} from '@angular/core/testing';
+
+import { SidebarComponent } from './sidebar.component';
+
+describe('Component: Sidebar', () => {
+  it('should create an instance', () => {
+    let component = new SidebarComponent();
+    expect(component).toBeTruthy();
+  });
+});